requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# PDF processing
pdfminer.six==20221105
PyMuPDF==1.23.8

# Template engine
jinja2==3.1.3

# Text processing
python-dateutil==2.8.2
regex==2023.12.25

# Optional: Search (lightweight)
whoosh==2.7.4

# Optional: NLP (CPU-only)
# sentence-transformers==2.3.1  # Uncomment if needed

# CLI
click==8.1.7
rich==13.7.0
